+++ src/main/java/br/com/fiap/Portaria/controller/MoradorController.java
@@ -2,13 +2,13 @@ package br.com.fiap.Portaria.controller;
 
 import br.com.fiap.Portaria.dto.MoradorRequestDTO;
 import br.com.fiap.Portaria.dto.MoradorResponseDTO;
-import br.com.fiap.Portaria.entity.Morador;
 import br.com.fiap.Portaria.service.MoradorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
 
 @RestController
 @RequestMapping("/moradores")
@@ -23,7 +23,7 @@ public class MoradorController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<MoradorResponseDTO> buscarPorId(@PathVariable Long id) {
+    public ResponseEntity<MoradorResponseDTO> buscarPorId(@PathVariable Integer id) {
         return moradorService.buscarPorId(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
@@ -35,7 +35,7 @@ public class MoradorController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<MoradorResponseDTO> atualizar(@PathVariable Long id, @RequestBody MoradorRequestDTO moradorRequestDTO) {
+    public ResponseEntity<MoradorResponseDTO> atualizar(@PathVariable Integer id, @RequestBody MoradorRequestDTO moradorRequestDTO) {
         try {
             return ResponseEntity.ok(moradorService.atualizar(id, moradorRequestDTO));
         } catch (RuntimeException e) {
@@ -44,7 +44,7 @@ public class MoradorController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletar(@PathVariable Long id) {
+    public ResponseEntity<Void> deletar(@PathVariable Integer id) {
         moradorService.deletar(id);
         return ResponseEntity.noContent().build();
     }
