+++ src/main/java/br/com/fiap/Portaria/repository/MoradorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MoradorRepository extends JpaRepository<Morador, Long> {
+public interface MoradorRepository extends JpaRepository<Morador, Integer> {
 }
